breaking out the code into proper files in src
